// app/types/access-management.ts
export type ServerStatus = 'online' | 'offline' | 'warning' | 'error'
export type ServerAuthType = 'password' | 'key' | 'local'

export interface ServerInfo {
  id: string
  name: string
  hostname: string
  ip: string
  port: number
  status: ServerStatus
  os: string
  version: string
  location: string
  tags: string[]
  description?: string
  username: string
  password?: string
  keyPath?: string
  authType: ServerAuthType
  isActive: boolean
  datacenter?: string
  lastConnectedAt?: Date | null
  userId: string
  createdAt: Date
  updatedAt: Date
}
// ...
export interface ServerFormData {
  name: string
  hostname: string
  ip: string
  port: number
  username: string
  password?: string
  keyPath?: string
  authType: ServerAuthType
  os?: string
  version?: string
  location?: string
  tags: string[]
  description?: string
}